# api/openapi.yaml
openapi: 3.0.3
info:
  title: NEX OSS/BSS ISP Cloud Platform API
  version: 3.4.0
  description: >
    Final pre-coding OpenAPI contract for NEX OSS/BSS v3.4 Router-Centric.
    Network topology rule: Customer -> Service -> Router -> Router Interface -> Radius NAS -> FreeRadius.
servers:
  - url: https://api.nexbil.local/api/v1
    description: Local/staging API base URL
security:
  - bearerAuth: []
...
paths:
...
  /tenants/{tenant_id}/customers/{customer_id}:
    get:
      tags: [Customers]
      summary: Get customer detail
      operationId: getCustomer
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
    patch:
      tags: [Customers]
      summary: Update customer
      operationId: updateCustomer
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
    delete:
      tags: [Customers]
      summary: Deactivate customer
      operationId: deleteCustomer
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '204':
          description: Customer deactivated
...
  /tenants/{tenant_id}/services/{service_id}:
    get:
      tags: [Services]
      summary: Get service detail
      operationId: getService
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Service detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
    patch:
      tags: [Services]
      summary: Update service data
      operationId: updateService
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: Service updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
  /tenants/{tenant_id}/services/{service_id}/activate:
    post:
      tags: [Services]
      summary: Activate service
      operationId: activateService
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Service activated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
        '409':
          description: Router mapping or Radius requirement incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /tenants/{tenant_id}/services/{service_id}/suspend:
    post:
      tags: [Services]
      summary: Suspend service and Radius user
      operationId: suspendService
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                invoice_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Service suspended
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
  /tenants/{tenant_id}/services/{service_id}/unsuspend:
    post:
      tags: [Services]
      summary: Unsuspend service and Radius user
      operationId: unsuspendService
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                payment_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Service unsuspended
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Service'
...
  /tenants/{tenant_id}/routers/{router_id}:
    get:
      tags: [Routers]
      summary: Get router detail
      operationId: getRouter
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/RouterId'
      responses:
        '200':
          description: Router detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Router'
    put:
      tags: [Routers]
      summary: Replace/update router
      operationId: updateRouter
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/RouterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouterUpdateRequest'
      responses:
        '200':
          description: Router updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Router'
...
  /tenants/{tenant_id}/router-script-generator:
    post:
      tags: [Routers]
      summary: Generate MikroTik RouterOS Radius script
      operationId: generateRouterScript
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouterScriptGenerateRequest'
      responses:
        '200':
          description: Script generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GeneratedScript'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenant_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CustomerId:
      name: customer_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ServiceId:
      name: service_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RouterId:
      name: router_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
        meta:
          type: object
          properties:
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time
    ErrorEnvelope:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          type: object
          properties:
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        external_ref:
          type: string
          nullable: true
        type:
          type: string
          enum: [individual, business]
        name:
          type: string
        billing_contact:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [prospect, lead, qualified, contracted, active, suspended, terminated]
...
    CustomerUpdateRequest:
      type: object
      properties:
        external_ref:
          type: string
        type:
          type: string
        name:
          type: string
        billing_contact:
          type: object
        status:
          type: string
...
    ServiceStatus:
      type: string
      enum: [requested, provisioning, active, suspended, terminated, failed]
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        service_category_id:
          type: string
          format: uuid
        contract_id:
          type: string
          format: uuid
          nullable: true
        cid:
          type: string
        status:
          $ref: '#/components/schemas/ServiceStatus'
        metadata:
          type: object
...
    ServiceUpdateRequest:
      type: object
      properties:
        cid:
          type: string
        metadata:
          type: object
...
    RouterRole:
      type: string
      enum: [core_router, aggregation_router, edge_router, pppoe_router, bng, wireless_gateway, pop_router, bts_router]
    RouterStatusValue:
      type: string
      enum: [draft, provisioning, active, warning, critical, maintenance, offline, retired]
    SnmpStatusValue:
      type: string
      enum: [reachable, unreachable, auth_failed, not_configured]
    Router:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        router_name:
          type: string
        hostname:
          type: string
        vendor:
          type: string
        model:
          type: string
        serial_number:
          type: string
        router_role:
          $ref: '#/components/schemas/RouterRole'
        site_name:
          type: string
        management_ip:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        status:
          $ref: '#/components/schemas/RouterStatusValue'
        snmp_status:
          $ref: '#/components/schemas/SnmpStatusValue'
    RouterCreateRequest:
      type: object
      required: [router_name, hostname, router_role, management_ip]
      properties:
        router_name:
          type: string
        hostname:
          type: string
        vendor:
          type: string
        model:
          type: string
        serial_number:
          type: string
        router_role:
          $ref: '#/components/schemas/RouterRole'
        site_name:
          type: string
        management_ip:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        snmp_profile:
          type: object
    RouterUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/RouterCreateRequest'
        - type: object
          properties:
            status:
              $ref: '#/components/schemas/RouterStatusValue'
            snmp_status:
              $ref: '#/components/schemas/SnmpStatusValue'
...
    RouterScriptGenerateRequest:
      type: object
      required: [router_id, os_version, script_type, radius_server_ip, radius_secret]
      properties:
        router_id:
          type: string
          format: uuid
        os_version:
          type: string
          enum: [ROS6, ROS7]
        script_type:
          type: string
          enum: [PPPoE, Hotspot]
        radius_server_ip:
          type: string
        radius_secret:
          type: string
        service_profile:
          type: string
    GeneratedScript:
      type: object
      properties:
        router_id:
          type: string
          format: uuid
        os_version:
          type: string
        script_type:
          type: string
        script:
          type: string
        warnings:
          type: array
          items:
            type: string
...