# apps/api-laravel/app/Models/Service.php
<?php

namespace App\Models;

use App\Models\Concerns\NexModel;
use App\Models\Traits\BelongsToTenant;
use Illuminate\Database\Eloquent\Relations\BelongsTo;
use Illuminate\Database\Eloquent\Relations\HasMany;

class Service extends NexModel
{
    use BelongsToTenant;

    protected function casts(): array
    {
        return ['metadata' => 'array', 'activated_at' => 'datetime', 'suspended_at' => 'datetime', 'terminated_at' => 'datetime'];
    }

    public function customer(): BelongsTo
    {
        return $this->belongsTo(Customer::class);
    }

    public function product(): BelongsTo
    {
        return $this->belongsTo(Product::class);
    }

    public function serviceCategory(): BelongsTo
    {
        return $this->belongsTo(ServiceCategory::class);
    }

    public function routerMappings(): HasMany
    {
        return $this->hasMany(ServiceRouterMapping::class);
    }

    public function radiusUsers(): HasMany
    {
        return $this->hasMany(RadiusUser::class);
    }
}